custom_exporter/requirements.txt
prometheus-client==0.20.0
psycopg2-binary==2.9.9

